Default Objective Function

A default objective function that combines adjacency and balance scores.

Usage

objective_function(
  adj_weight = getOption("speed.adj_weight", 1),
  bal_weight = getOption("speed.bal_weight", 1)
)

Arguments

adj_weight

Weight for adjacency score (default: 1)

bal_weight

Weight for balance score (default: 1)

Value

A function which returns numeric value representing the score of the design (lower is better) with a signature function(design_matrix, layout_df, swap, spatial_cols). See signature details in objective_function_signature.

Examples

design_matrix <- matrix(c(1, 2, 2, 1, 3, 3, 1, 3, 3), nrow = 3, ncol = 3)
layout_df <- data.frame(
  row = rep(1:3, each = 3),
  col = rep(1:3, times = 3)
)
objective_function()(design_matrix, layout_df, "treatment", c("row", "col"))
#> [1] 15

# create an objective function including even distribution and neighbor balance introduced by Piepho 2018
objective_function_piepho <- function(design_matrix, layout_df, swap, spatial_cols) {
  ed <- calculate_ed(design_matrix)
  ed_score <- -sum(unlist(lapply(ed, function(ed_rep) ed_rep$min_mst)))
  nb_score <- calculate_nb(design_matrix)$max_nb

  adj_bal_score <- objective_function()(design_matrix, layout_df, swap, spatial_cols)

  return(nb_score + ed_score + adj_bal_score)
}
objective_function_piepho(design_matrix, layout_df, "treatment", c("row", "col"))
#> [1] 13
# usage in speed, speed(..., obj_function = objective_function_piepho)